feat: Disable root motion for selected AnimSequence assets
This code change adds functionality to disable root motion for selected AnimSequence assets. It iterates through the selected assets and checks if root motion is enabled. If so, it disables root motion for the asset. The names of the assets for which root motion was disabled are displayed in the dialog box.