refactor: Remove shader root directory forcefully
This code change removes the shader root directory and its contents forcefully. The function `remove_dir_force()` is added to clear the readonly bit and reattempt the removal of the directory. The `shutil.rmtree()` function is used to remove the directory and its contents. This change ensures that the shader root directory is completely removed.